+++ RELEASE-NOTES.md
@@ -3,7 +3,7 @@
 ## 2020-10-02 - v0.4.2
 
 - AB#27: (bugfix) Redirect to application only with 'login' token.
-
+- AB#28: (Support) `cancel url`.
 
 ## 2020-09-28 - v0.4.1